Update tests to use test data created on dev stack.
A number of fixes were added and functions created to clean the data so that comparisons between actual and expected results can be made.

The README has been updated so that users will know what steps in docker ckan stack are needed to set it up for the tests to pass.